+++ src/euroeval/data_models.py
@@ -885,7 +885,7 @@ class BenchmarkResult(pydantic.BaseModel):
             results_path:
                 The path to the results file.
         """
-        json_str = json.dumps(self.model_dump())
+        json_str = json.dumps(self.model_dump(), ensure_ascii=False)
         with results_path.open("a") as f:
             f.write("\n" + json_str)
 
@@ -968,14 +968,28 @@ class GenerativeModelOutput:
     Attributes:
         sequences:
             The generated sequences.
-        scores:
+        predicted_labels (optional):
+            The predicted labels from the `sequences` and sometimes also `scores`. Can
+            be None if the labels have not been predicted yet. Defaults to None.
+        scores (optional):
             The scores of the sequences. This is an array of shape (batch_size,
             num_tokens, num_logprobs, 2), where the last dimension contains the
-            token and its logprob. Can be None if the scores are not available.
+            token and its logprob. Can be None if the scores are not available. Defaults
+            to None.
+        metadatas (optional):
+            All the metadata fields for the samples, including ground truth labels (if
+            applicable). Defaults to an empty list.
     """
 
     sequences: c.Sequence[str]
+    predicted_labels: c.Sequence[str] | None = None
     scores: c.Sequence[c.Sequence[c.Sequence[tuple[str, float]]]] | None = None
+    metadatas: list["HashableDict | None"] = field(default_factory=list)
+
+    def __post_init__(self) -> None:
+        """Post-initialisation."""
+        if not self.metadatas:
+            self.metadatas = [None] * len(self.sequences)
 
 
 @dataclass
@@ -985,14 +999,22 @@ class SingleGenerativeModelOutput:
     Attributes:
         sequence:
             The generated sequence.
-        scores:
+        predicted_label (optional):
+            The predicted label from the `sequence` and sometimes also `scores`. Can be
+            None if the label has not been predicted yet. Defaults to None.
+        scores (optional):
             The scores of the sequence. This is an array of shape (num_tokens,
             num_logprobs, 2), where the last dimension contains the token and its
-            logprob. Can be None if the scores are not available.
+            logprob. Can be None if the scores are not available. Defaults to None.
+        metadata (optional):
+            The metadata fields for the sample, including ground truth labels (if
+            applicable). Can be None if the metadata is not available. Defaults to None.
     """
 
     sequence: str
+    predicted_label: str | None = None
     scores: c.Sequence[c.Sequence[tuple[str, float]]] | None = None
+    metadata: "HashableDict | None" = None
 
 
 @dataclass