+++ CHANGELOG.md
@@ -9,6 +9,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 
+- We now add all metadata (including ground truth labels, if applicable) to the model
+  cache when debug mode is enabled (with `--debug` or `debug=True`). We have added a
+  [section in the
+  documentation](https://euroeval.com/python-package/#analysing-the-results) on how to
+  use this feature.
 - When evaluating community evaluation datasets from the Hugging Face Hub, we now
   require that the user actively set `--trust-remote-code` (or `trust_remote_code=True`
   if running with the `Benchmarker` class). This is to prevent accidental execution of